requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
python-pptx==1.0.2
openai==1.99.9
azure-cognitiveservices-speech==1.45.0
python-dotenv==1.1.1
jinja2==3.1.6
pywin32==310
comtypes==1.4.13
